BAU - Move UnsuccesfulCredentialResponseException to shared and make it a checked Exception
- Move this to shared so we can also use it in the IPV lambdas
- Make this a checked exception as we want to (and already do) specific behaviour when it is thrown such as returning the user to the frontend error page
